+++ NEWS.md
@@ -1,3 +1,17 @@
+# rqti 0.2.1
+
+## Bug fixes
+
+* In metadata default value for the slot contribution date is assigned to 
+Date(0). 
+
+## New features
+
+* method createQtiTask() can create XML or zip files, to control this use 
+parameter zip=TRUE.
+
+* QTIJS rendering shows feedback and score for tasks.
+
 # rqti 0.2.0
 
 ## Bug fixes